Moves tests next to files
