+++ src/components/CampaignSummary.test.tsx
@@ -6,7 +6,7 @@ import {
   wrapInTestApp,
 } from "@backstage/test-utils";
 import { configApiRef } from "@backstage/core-plugin-api";
-import CampaignSummary from "../components/CampaignSummary";
+import CampaignSummary from "./CampaignSummary";
 import { ChecksByCampaign } from "../types/OpsLevelData";
 
 const getMockConfig = () =>
