+++ src/components/OverallMaturityOverview.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import OverallMaturityOverview from "../components/OverallMaturityOverview";
+import OverallMaturityOverview from "./OverallMaturityOverview";
 
 describe("OverallMaturityOverview", () => {
   it("renders a progress bar in loading mode", () => {
